+++ shared/constants.go
@@ -1,0 +1,6 @@
+package shared
+
+const (
+	MetricProxiedSuccess string = "proxied_success"
+	MetricProxiedErrors  string = "proxied_errors"
+)